myblog/requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.2.1
django-crispy-forms==2.0
django-environ==0.10.0
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
sqlparse==0.4.4
tomli==2.0.1
